test suit built for ecommerce type cluster
