+++ 11/topic-ex/topic-reciever3.js
@@ -1,14 +1,13 @@
 const amqp = require('amqplib/callback_api');
 const dotenv = require('dotenv').config();
-const uri = process.env.URI
-
+// const uri = process.env.URI
+const uri = 'amqp://test:test@192.168.0.236:5672'
 
 const script = {}
 
 script.run = () => {
-  const ex = 'direct'
+  const ex = 'topic'
   const queue = 'inventory' //queue name
-  
   
   amqp.connect(uri, (err, conn) => {
     conn.createChannel((err, ch) => {
@@ -16,7 +15,6 @@ script.run = () => {
         ch.consume(queue, function(msg) {
         console.log(" .-. >>", msg.content.toString());
     
-  
       })
       // ch.assertQueue(q, {durable: true}); // not durable/presistent message
