--- consumer/rpc_client.js
@@ -1,40 +1,0 @@
-#!/usr/bin/env node
-
-var amqp = require('amqplib/callback_api');
-
-const URI = 'amqp://test:test@192.168.0.236:5672'
-
-var args = process.argv.slice(2);
-
-if (args.length == 0) {
-  console.log("Usage: rpc_client.js num");
-  process.exit(1);
-}
-
-amqp.connect(URI, function(err, conn) {
-  conn.createChannel(function(err, ch) {
-    ch.assertQueue('', {exclusive: true}, function(err, q) {
-      var corr = generateUuid();
-      var num = parseInt(args[0]);
-
-      console.log(' [x] Requesting fib(%d)', num);
-
-      ch.consume(q.queue, function(msg) {
-        if (msg.properties.correlationId == corr) {
-          console.log(' [.] Got %s', msg.content.toString());
-          setTimeout(function() { conn.close(); process.exit(0) }, 500);
-        }
-      }, {noAck: true});
-
-      ch.sendToQueue('rpc_queue',
-      new Buffer.from(num.toString()),
-      { correlationId: corr, replyTo: q.queue });
-    });
-  });
-});
-
-function generateUuid() {
-  return Math.random().toString() +
-         Math.random().toString() +
-         Math.random().toString();
-}
