+++ ecommerce/default-ex/default-receiver.js
@@ -4,15 +4,14 @@ const uri = process.env.URI
 
 
 //Receive messages on default exchange
+const script = {}
+script.run = () => {
+q = 'items_page' //queue 
 
-ex = 'amq.direct'
-q = 'cyan' //queue name
-
-
-amqp.connect(uri, (err, conn) => {
+module.exports = amqp.connect(uri, (err, conn) => {
   conn.createChannel((err, ch) => {
 
-    ch.assertQueue(q, {durable: true}); // not durable/presistent message
+    ch.assertQueue(q, {durable: false}); // not durable/presistent message
 
     console.log(`[*] listening on queue: ${q}`);
     ch.consume(q, function(msg) {
@@ -20,3 +19,5 @@ amqp.connect(uri, (err, conn) => {
 }, {noAck: true}); // no acknowledge means messages will "fire and forget"
   });
 });
+}
+module.exports = script;
