+++ ecommerce/fanout-ex/fanout-receiver.js
@@ -5,14 +5,13 @@ const uri = 'amqp://test:test@192.168.0.236:5672'
 
 const script = {}
 
-script.connectToRabbit = () => { 
+script.run = () => { 
   amqp.connect(uri, (err, conn) => {
     conn.createChannel((err, ch) => {
   
       const ex = 'amq.fanout';
       
-      ch.assertExchange(ex, 'fanout', {durable: true});
-      ch.assertQueue('email', {exclusive: false}, (err, q)=> {
+      ch.assertQueue('email', {durable: false}, (err, q)=> {
         console.log('.-. waiting for messages', q.queue);
         
         //arguments: queueName  - exchangeName - key
@@ -33,7 +32,7 @@ script.connectToRabbit = () => {
     conn.createChannel((err, ch) => {
       const ex = 'amq.fanout';
   
-    ch.assertQueue('inventory', {exclusive: false}, (err, q)=> {
+    ch.assertQueue('inventory', {durable: false}, (err, q)=> {
       console.log('.-. waiting for messages', q.queue);
       
       //arguments: queueName  - exchangeName - key
@@ -49,7 +48,7 @@ script.connectToRabbit = () => {
     conn.createChannel((err, ch) => {
       const ex = 'amq.fanout';
   
-     ch.assertQueue('accounting', {exclusive: false}, (err, q)=> {
+     ch.assertQueue('accounting', {durable: false}, (err, q)=> {
         console.log('.-. waiting for messages', q.queue);
         
         //arguments: queueName  - exchangeName - key
@@ -60,6 +59,23 @@ script.connectToRabbit = () => {
           } 
         }, {noAck: true})
     })
+  })
+
+
+  conn.createChannel((err, ch) => {
+    const ex = 'amq.fanout';
+
+  ch.assertQueue('shipping', {durable: false}, (err, q)=> {
+    console.log('.-. waiting for messages', q.queue);
+    
+    //arguments: queueName  - exchangeName - key
+    ch.bindQueue('shipping', ex, 'successful_txn');
+    ch.consume(q.queue, function(msg){
+      if (msg.content) {
+        console.log('message:', msg.content.toString())
+      } 
+    }, {noAck: true})
+  })
   })
   })
