+++ ecommerce/topic-ex/topic-receiver2.js
@@ -1,6 +1,6 @@
 const amqp = require('amqplib/callback_api');
 const dotenv = require('dotenv').config();
-const uri = process.env.URI
+const uri = 'amqp://test:test@192.168.0.236:5672'
 
 
 //Receive messages on default exchange
@@ -9,12 +9,18 @@ const ex = 'amq.topic'
 const queue = 'inventory' //queue name
 
 
+
 amqp.connect(uri, (err, conn) => {
   conn.createChannel((err, ch) => {
-      ch.bindQueue(queue, ex, '*.c')
-      ch.consume(queue, function(msg) {
+    ch.assertQueue(queue, {durable: false}, function(err, q) {
+      if (err) {
+        console.log(err)
+      }
+      console.log(`[*] listening on queue: ${q}`);
+      ch.bindQueue(q.queue, ex, '*.c')
+      ch.consume(q.queue, function(msg) {
       console.log(" .-. >>", msg.content.toString());
-  
+  }, {noAck: true});
 
     })
     // ch.assertQueue(q, {durable: true}); // not durable/presistent message
